requirements.txt
whoosh==2.7.4
langchain==0.2.16
langchain-core==0.2.39
langchain-community==0.2.17
langchain-text-splitters==0.2.2
chromadb==0.5.5
pypdf==4.2.0
fastapi==0.111.0
uvicorn==0.30.6
python-dotenv==1.0.1
requests==2.32.3
typing-extensions==4.12.2
cachetools==5.3.3
langchain_groq
sentence-transformers
transformers